Initial commit of clone_scanner.py and my weechat_scripts repo.
My personal repository of scripts for the WeeChat chat client,
www.weechat.org Scripts I've written to add features to WeeChat I felt
were missing.  clone_scanner.py is a script to detect if a joining user
is already on the channel from a different nick with the same hostname.